# Conflicts:
#	Assets/Scripts/Data/Augmenter_1.asset
#	Assets/Scripts/Data/Augmenter_2.asset
#	Assets/Scripts/Data/Augmenter_3.asset
#	Assets/Scripts/Data/Augmenter_4.asset
#	Assets/Scripts/Data/Augmenter_6.asset

+++ Assets/Scripts/Data/Augmenter_Data.cs
@@ -19,7 +19,7 @@ public class Augmenter_Data : ScriptableObject
     [Header("# Ctrl_Num")]
     //public int num;
     public int count;
-    public List<int> num;
+    public List<float> num;
 
     
 }
